% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Gauss_Newton_method.R
\name{Cluster_Gauss_Newton_EBE_method}
\alias{Cluster_Gauss_Newton_EBE_method}
\title{Cluster_Gauss_Newton_EBE_method}
\usage{
Cluster_Gauss_Newton_EBE_method(
  CGNM_result,
  nonlinearFunction,
  individualIndices_vec,
  numRepeat = 1,
  keepInitialDistribution = NA,
  algorithmParameter_EBEweight = 1,
  ...
)
}
\arguments{
\item{CGNM_result}{(required input) \emph{A list} stores the computational result from Cluster_Gauss_Newton_method() function in CGNM package.}

\item{nonlinearFunction}{(required input) \emph{A function with input of a vector x of real number of length n and output a vector y of real number of length m.} In the context of model fitting the nonlinearFunction is \strong{the model}.  Given the CGNM does not assume the uniqueness of the minimizer, m can be less than n.  Also CGNM does not assume any particular form of the nonlinear function and also does not require the function to be continuously differentiable (see Appendix D of our publication for an example when this function is discontinuous).}

\item{individualIndices_vec}{(required input) \emph{A string vector} where the each element corresponds to each observation and the value of the element is the ID that will be used for EBE. Must be the same as the length of the target vector.}

\item{numRepeat}{(default: 1) \emph{A vector of integers} number of EBE per individual to obtain.}

\item{keepInitialDistribution}{(default: NA) \emph{A vector of TRUE or FALSE} of length n User can specify if the initial distribution of one of the input variable (e.g. parameter) to be kept as the initial iterate throughout CGNM iterations.}

\item{algorithmParameter_EBEweight}{(default: 9) \emph{A number}}

\item{...}{Further arguments to be supplied to nonlinearFunction}
}
\value{
list of a matrix X, Y,residual_history, initialX, bootstrapX, bootstrapY as well as a list runSetting.
\enumerate{\item X, Y, residual_history, initialX: identical to what was given as CGNM_result.
\item X: \emph{a num_bootstrapSample by n matrix} which stores the the X values that was sampled using residual resampling bootstrap analyses (In terms of model fitting this is the parameter combinations with variabilities that represent \strong{parameter estimation uncertainties}.).
\item Y: \emph{a num_bootstrapSample by m matrix} which stores the nonlinearFunction evaluated at the corresponding bootstrap analyses results in matrix bootstrapX above. In the context of model fitting each row corresponds to \strong{the model simulations}.
\item runSetting: identical to what is given as CGNM_result but in addition including num_bootstrapSample and indicesToUseAsInitialIterates.}
}
\description{
obtain ebe empirical bayes estimate (EBE) using CGNM.
}
\examples{
##lip-flop kinetics (an example known to have two distinct solutions)

model_analytic_function=function(x){

 observation_time=c(0.1,0.2,0.4,0.6,1,2,3,6,12)
 Dose=1000
 F=1

 ka=x[1]
 V1=x[2]
 CL_2=x[3]
 t=observation_time

 Cp=ka*F*Dose/(V1*(ka-CL_2/V1))*(exp(-CL_2/V1*t)-exp(-ka*t))

 log10(Cp)
}

observation=log10(c(4.91, 8.65, 12.4, 18.7, 24.3, 24.5, 18.4, 4.66, 0.238))

CGNM_result=Cluster_Gauss_Newton_method(
nonlinearFunction=model_analytic_function,
targetVector = observation, num_iteration = 10, num_minimizersToFind = 100,
initial_lowerRange = c(0.1,0.1,0.1), initial_upperRange =  c(10,10,10),
lowerBound=rep(0,3), ParameterNames=c("Ka","V1","CL_2"), saveLog = FALSE)

CGNM_EBE=Cluster_Gauss_Newton_EBE_method(CGNM_result,
     nonlinearFunction=model_analytic_function, individualIndices_vec=seq(1,9))


}
