% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Gauss_Newton_method.R
\name{col_quantile}
\alias{col_quantile}
\title{col_quantile}
\usage{
col_quantile(data_in, prob)
}
\arguments{
\item{data_in}{(required input) \emph{a matrix or a data.grame} where the column-wise quantile wishes to be determined.}

\item{prob}{(required input) \emph{a number} quantile expressed as in the probability.}
}
\value{
\emph{a vector of number dim(data_in)[2]} containing: quantile of the each column where the probability is specified as "prob"
}
\description{
Obtain columb wise quantile
}
\examples{
A=matrix(seq(1,100),nrow = 25)
col_quantile(A, 0.5)
}
