% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCGNM_related.R
\name{make_ShinyCGNM_doseData}
\alias{make_ShinyCGNM_doseData}
\title{make_ShinyCGNM_doseData}
\usage{
make_ShinyCGNM_doseData(ID, dose, dosing.to, start.time, rate, fileName = NA)
}
\arguments{
\item{ID}{(required input) \emph{string or vector}}

\item{dose}{(required input) \emph{number or numeric vector}}

\item{dosing.to}{(required input) \emph{string or string vector}}

\item{start.time}{(required input) \emph{number or numeric vector}}

\item{rate}{(required input) \emph{NA, number or numeric vector} set ti BA uf bolus}

\item{fileName}{(default: NA) \emph{NA or string}}
}
\value{
\emph{data.frame} if fileName is NA \emph{Null} if fileName is not NA but instead write out the csv file
}
\description{
A helper function to write out the csv file that can be read in as the dose file in shinyCGNM
}
\examples{

make_ShinyCGNM_doseData(
ID=seq(1,5),
dose=10,
dosing.to="A_admin",
start.time=0,
rate=NA
)

}
