% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Gauss_Newton_method.R
\name{shinyCGNM}
\alias{shinyCGNM}
\title{shinyCGNM}
\usage{
shinyCGNM()
}
\value{
NULL and start graphifcal user interface.
}
\description{
Start Shiny app that assist the user to make R-script to conduct PBPK model fitting using CGNM.
}
\examples{
\dontrun{
shinyCGNM()
}
}
